+++ lib/graph/vertex.dart
@@ -3,7 +3,7 @@
 import 'package:flutteraji/chaturaji/move.dart';
 
 class Vertex {
-  List<int> Q = [0, 0, 0, 0];
+  List<double> Q = [0.0, 0.0, 0.0, 0.0];
   int N = 0;
   Map<Move, int> edges = {};
